chore: move logheader functions in separate file
